backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
motor==3.6.0
pymongo<4.10,>=4.9
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0
